refactor architecture header now contains sunrise/sunset info and header appears on all screens
